requirements.txt
flask==2.0.1
werkzeug==2.0.3
openai==1.3.0
python-dotenv==0.19.0
flask-cors==3.0.10
